ICMSLST-479 Add sanctions and adhoc license & refactor of firearms.
Also:

* Add sanctions and adhoc license views.
* Refactor firearms to be more generic.
* Add models to admin.py.
