+++ web/domains/case/_import/sanctions/models.py
@@ -1,0 +1,6 @@
+from ..models import ImportApplication
+
+
+class SanctionsAndAdhocApplication(ImportApplication):
+    PROCESS_TYPE = "SanctionsAndAdhocApplication"
+    pass
