+++ web/templates/internal.html
@@ -14,3 +14,4 @@
     {% endblock %}
   </div>
 {% endblock %}
+
