+++ web/templates/partial/case/import/sanctions/sidebar.html
@@ -1,0 +1,14 @@
+<ul class="menu-out">
+  <li>
+      <a href="{{ url('import:sanctions-applicant-details', kwargs={'pk': process.pk}) }}"><b>Applicant Details</b></a>
+  </li>
+  <li>
+    <a href="{{ url('import:sanctions-and-adhoc-licence-application-details', kwargs={'pk': process.pk}) }}">Sanctions and Adhoc Licence Application Details</a>
+  </li>
+  <li>
+    <a href="{{ url('import:sanctions-validation-summary', kwargs={'pk': process.pk}) }}">Validation Summary</a>
+  </li>
+  <li>
+    <a href="{{ url('import:sanctions-submit', kwargs={'pk': process.pk}) }}">Submit</a>
+  </li>
+</ul>
