+++ web/templates/partial/case/import/sidebar.html
@@ -1,12 +1,18 @@
-<h5>Firearms and Ammunition</h5>
+<h5>Sanctions and Adhoc</h5>
+
 <ul class="menu-out">
+  <li>
+    <a href="{{ url('import:create-sanctions') }}">Sanctions and Adhoc License Application</a>
+  </li>
+    <h5>Firearms and Ammunition</h5>
   <li>
     <a href="#">Deactivated Firearms Licence</a>
   </li>
   <li>
-    <a href="{{ url('create-oil') }}">Open Individual Licence (OIL)</a>
+    <a href="{{ url('import:create-oil') }}">Open Individual Licence (OIL)</a>
   </li>
   <li>
     <a href="#">Specific Individual Licence (SIL)</a>
   </li>
+
 </ul>
