+++ web/templates/web/domains/case/import/create.html
@@ -1,0 +1,46 @@
+{% extends "layout/sidebar.html" %}
+{% import "forms/forms.html" as forms %}
+{% import "forms/fields.html" as fields %}
+
+{% block css %}
+  {{ super() }}
+  {{ form.media.css }}
+{% endblock %}
+
+{% block sidebar %}
+    {% include "partial/case/import/sidebar.html" %}
+{% endblock %}
+
+{% block main_content %}
+  {% call forms.form(method='post', csrf_input=csrf_input) -%}
+    <h3>{{ import_application_type.get_type_description() }}</h3>
+
+    {% for field in form %}
+      {{fields.field(field)}}
+    {% endfor %}
+
+    <div class="container">
+      <div class="row">
+        <div class="three columns"></div>
+        <div class="eight columns">
+          <ul class="menu-out flow-across">
+            <li>
+              <input
+                type="submit"
+                name="action"
+                class="primary-button button"
+                value="Create" />
+            </li>
+          </ul>
+        </div>
+      </div>
+    </div>
+
+  {%endcall %}
+{% endblock %}
+
+{% block page_js %}
+  {{ super() }}
+  {{ form.media.js }}
+{% endblock %}
+
