+++ web/templates/web/domains/case/import/sanctions/sanctions_validation_summary.html
@@ -1,0 +1,10 @@
+{% extends "flow/task-base.html" %}
+
+
+{% block sidebar %}
+    {% include "partial/case/import/sanctions/sidebar.html" %}
+{% endblock %}
+
+{% block main_content %}
+
+{% endblock main_content %}
